Refactor: Login form and form test
